+++ document/qhexcursor.cpp
@@ -44,20 +44,20 @@ const QHexPosition &QHexCursor::selectionEnd() const
 const QHexPosition &QHexCursor::position() const { return m_position; }
 QHexCursor::InsertionMode QHexCursor::insertionMode() const { return m_insertionmode; }
 int QHexCursor::selectionLength() const { return this->selectionEnd() - this->selectionStart() + 1;  }
-int QHexCursor::currentLine() const { return m_position.line; }
+quint64 QHexCursor::currentLine() const { return m_position.line; }
 int QHexCursor::currentColumn() const { return m_position.column; }
 int QHexCursor::currentNibble() const { return m_position.nibbleindex; }
-int QHexCursor::selectionLine() const { return m_selection.line; }
+quint64 QHexCursor::selectionLine() const { return m_selection.line; }
 int QHexCursor::selectionColumn() const { return m_selection.column; }
 int QHexCursor::selectionNibble() const { return m_selection.nibbleindex;  }
 
-bool QHexCursor::isLineSelected(int line) const
+bool QHexCursor::isLineSelected(quint64 line) const
 {
     if(!this->hasSelection())
         return false;
 
-    int first = std::min(m_position.line, m_selection.line);
-    int last = std::max(m_position.line, m_selection.line);
+    quint64 first = std::min(m_position.line, m_selection.line);
+    quint64 last = std::max(m_position.line, m_selection.line);
 
     if((line < first) || (line > last))
         return false;
@@ -76,7 +76,7 @@ void QHexCursor::clearSelection()
 void QHexCursor::moveTo(const QHexPosition &pos) { this->moveTo(pos.line, pos.column, pos.nibbleindex); }
 void QHexCursor::select(const QHexPosition &pos) { this->select(pos.line, pos.column, pos.nibbleindex); }
 
-void QHexCursor::moveTo(int line, int column, int nibbleindex)
+void QHexCursor::moveTo(quint64 line, int column, int nibbleindex)
 {
     m_selection.line = line;
     m_selection.column = column;
@@ -85,7 +85,7 @@ void QHexCursor::moveTo(int line, int column, int nibbleindex)
     this->select(line, column, nibbleindex);
 }
 
-void QHexCursor::select(int line, int column, int nibbleindex)
+void QHexCursor::select(quint64 line, int column, int nibbleindex)
 {
     m_position.line = line;
     m_position.column = column;
@@ -94,7 +94,7 @@ void QHexCursor::select(int line, int column, int nibbleindex)
     emit positionChanged();
 }
 
-void QHexCursor::moveTo(int offset)
+void QHexCursor::moveTo(qint64 offset)
 {
     int line = offset / m_lineWidth;
     this->moveTo(line, offset - (line * m_lineWidth));
@@ -102,7 +102,7 @@ void QHexCursor::moveTo(int offset)
 
 void QHexCursor::select(int length) { this->select(m_position.line, std::min(m_lineWidth - 1, m_position.column + length - 1)); }
 
-void QHexCursor::selectOffset(int offset, int length)
+void QHexCursor::selectOffset(qint64 offset, int length)
 {
     this->moveTo(offset);
     this->select(length);
@@ -117,7 +117,7 @@ void QHexCursor::setInsertionMode(QHexCursor::InsertionMode mode)
 		emit insertionModeChanged();
 }
 
-void QHexCursor::setLineWidth(int width)
+void QHexCursor::setLineWidth(quint8 width)
 {
     m_lineWidth = width;
     m_position.lineWidth = width;
